Fix various timezone issues.
This fixes UTC offsets and abbreviations and UTC-local correspondence,
but keeps the meeting time at its old local time for North America and
Europe (shifting in Japan and Australia), at least for this week.  Add
an agenda item on whether we want to move an hour earlier.  (All
discussed with @dka before editing.)